pyproject.toml
[tool.poetry]
name = "concrete-biopython"
version = "0.0.1"
description = ""
authors = ["Lcressot <Lcressot@gmail.com>"]
readme = "README.md"
packages = [{include = "concrete_biopython"}]

[tool.poetry.dependencies]
python = ">=3.9, <3.12"
biopython = "^1.81"
ipykernel = "^6.23.1"
matplotlib = "^3.7.1"
concrete-python = "^2.2.0"
numpy = "^1.26.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
